Fix DPD Values to be signed

+++ Dev4Agriculture.ISO11783.ISOXML.Test/TaskDataAnalysisTests.cs
@@ -17,11 +17,11 @@ namespace Dev4Agriculture.ISO11783.ISOXML.Test
 
             //Testing Task Total
             Assert.IsTrue(isoxml.Data.Task[0].TryGetTotalValue(0x78, 0, out var totalInEffectiveTime, TLGTotalAlgorithmType.NO_RESETS));
-            Assert.AreEqual(totalInEffectiveTime, (uint)4531 /*Close to 75.5 minutes*/);
+            Assert.AreEqual(totalInEffectiveTime, 4531 /*Close to 75.5 minutes*/);
 
             //Testing Task Maximum
             Assert.IsTrue(isoxml.Data.Task[0].TryGetMaximum(0x43, 0, out var maximum));
-            Assert.AreEqual(maximum, (uint)12000);
+            Assert.AreEqual(maximum, 12000);
         }
 
         [TestMethod]
